backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
pandas==2.1.3
faker==20.1.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.8.3
python-dotenv==1.0.0
websockets==12.0
secure-smtplib==0.1.1  # For dummy email sending